requirements.txt
streamlit
torch
torchvision
pandas
Pillow
sklearn
seaborn
matplotlib
numpy
